moves job data to layout
