+++ frontend/app/jobs/[id]/layout.tsx
@@ -1,15 +1,36 @@
 'use client';
 import { SidebarNav } from '@/components/SideBarNav';
-import { Separator } from '@/components/ui/separator';
-import { useParams } from 'next/navigation';
+import SubPageHeader from '@/components/headers/SubPageHeader';
+import SkeletonForm from '@/components/skeleton/SkeletonForm';
+import { PageProps } from '@/components/types';
+import { Alert, AlertTitle } from '@/components/ui/alert';
+import { Button } from '@/components/ui/button';
+import { toast } from '@/components/ui/use-toast';
+import { useGetJob } from '@/libs/hooks/useGetJob';
+import { getErrorMessage } from '@/util/util';
+import { useRouter } from 'next/navigation';
 
-interface SettingsLayoutProps {
-  children: React.ReactNode;
-}
+export default function SettingsLayout({ children, params }: PageProps) {
+  const id = params?.id ?? '';
+  const basePath = `/jobs/${params?.id}`;
+  const { data, isLoading, mutate } = useGetJob(id);
+  const router = useRouter();
 
-export default function SettingsLayout({ children }: SettingsLayoutProps) {
-  const params = useParams();
-  const basePath = `/jobs/${params.id}`;
+  async function onTriggerJobRun(): Promise<void> {
+    try {
+      await triggerJobRun(id);
+      toast({
+        title: 'Job run triggered successfully!',
+      });
+      router.push(`/runs`);
+    } catch (err) {
+      console.error(err);
+      toast({
+        title: 'Unable to trigger job run',
+        description: getErrorMessage(err),
+      });
+    }
+  }
 
   const sidebarNavItems = [
     {
@@ -26,15 +47,36 @@ export default function SettingsLayout({ children }: SettingsLayoutProps) {
     },
   ];
 
-  return (
-    <div className=" space-y-6 p-10 pb-16 md:block">
-      <div className="space-y-0.5">
-        <h2 className="text-2xl font-bold tracking-tight">Job Overview</h2>
-        <p className="text-muted-foreground">
-          View and manage job configuration.
-        </p>
+  if (isLoading) {
+    return (
+      <div>
+        <SkeletonForm />
       </div>
-      <Separator className="my-6" />
+    );
+  }
+
+  if (!data?.job) {
+    return (
+      <div className="mt-10">
+        <Alert variant="destructive">
+          <AlertTitle>{`Error: Unable to retrieve job`}</AlertTitle>
+        </Alert>
+      </div>
+    );
+  }
+
+  return (
+    <div className="space-y-6 p-10 pb-16 md:block">
+      <div className="space-y-1">
+        <h2 className="text-2xl font-bold tracking-tight">Job Overview</h2>
+        <SubPageHeader
+          header={data?.job?.name || ''}
+          description={data?.job?.id || ''}
+          extraHeading={
+            <Button onClick={() => onTriggerJobRun()}>Trigger Run</Button>
+          }
+        />
+      </div>
       <div className="flex flex-col space-y-8 lg:flex-row lg:space-x-12 lg:space-y-0">
         <aside className="-mx-4 lg:w-[200px]">
           <SidebarNav items={sidebarNavItems} />
@@ -43,4 +85,16 @@ export default function SettingsLayout({ children }: SettingsLayoutProps) {
       </div>
     </div>
   );
+}
+
+async function triggerJobRun(jobId: string): Promise<void> {
+  const res = await fetch(`/api/jobs/${jobId}/create-run`, {
+    method: 'POST',
+    body: JSON.stringify({ jobId }),
+  });
+  if (!res.ok) {
+    const body = await res.json();
+    throw new Error(body.message);
+  }
+  await res.json();
 }
