+++ frontend/app/jobs/[id]/page.tsx
@@ -2,14 +2,9 @@
 import { PageProps } from '@/components/types';
 import { useGetJob } from '@/libs/hooks/useGetJob';
 
-import SubPageHeader from '@/components/headers/SubPageHeader';
 import SkeletonForm from '@/components/skeleton/SkeletonForm';
 import { Alert, AlertTitle } from '@/components/ui/alert';
-import { Button } from '@/components/ui/button';
-import { useToast } from '@/components/ui/use-toast';
 import { useGetJobStatus } from '@/libs/hooks/useGetJobStatus';
-import { getErrorMessage } from '@/util/util';
-import { useRouter } from 'next/navigation';
 import { ReactElement } from 'react';
 import JobPauseSwitch from './components/PauseSwitch';
 import JobRecentRuns from './components/RecentRuns';
@@ -19,24 +14,6 @@ export default function Page({ params }: PageProps): ReactElement {
   const id = params?.id ?? '';
   const { data, isLoading, mutate } = useGetJob(id);
   const { data: jobStatus } = useGetJobStatus(id);
-  const router = useRouter();
-  const { toast } = useToast();
-
-  async function onTriggerJobRun(): Promise<void> {
-    try {
-      await triggerJobRun(id);
-      toast({
-        title: 'Job run triggered successfully!',
-      });
-      router.push(`/runs`);
-    } catch (err) {
-      console.error(err);
-      toast({
-        title: 'Unable to trigger job run',
-        description: getErrorMessage(err),
-      });
-    }
-  }
 
   if (isLoading) {
     return (
@@ -58,13 +35,6 @@ export default function Page({ params }: PageProps): ReactElement {
 
   return (
     <div className="job-details-container">
-      <SubPageHeader
-        header={data?.job?.name || ''}
-        description={data?.job?.id || ''}
-        extraHeading={
-          <Button onClick={() => onTriggerJobRun()}>Trigger Run</Button>
-        }
-      />
       <div className="space-y-10">
         <JobPauseSwitch jobId={id} status={jobStatus?.status} mutate={mutate} />
         <JobScheduleCard job={data?.job} mutate={mutate} />
@@ -76,16 +46,4 @@ export default function Page({ params }: PageProps): ReactElement {
       </div>
     </div>
   );
-}
-
-async function triggerJobRun(jobId: string): Promise<void> {
-  const res = await fetch(`/api/jobs/${jobId}/create-run`, {
-    method: 'POST',
-    body: JSON.stringify({ jobId }),
-  });
-  if (!res.ok) {
-    const body = await res.json();
-    throw new Error(body.message);
-  }
-  await res.json();
 }
